+++ rumqttc-v4/src/eventloop.rs
@@ -34,6 +34,7 @@ use {
     crate::websockets::WsAdapter,
     crate::websockets::{UrlError, split_url, validate_response_headers},
     async_tungstenite::tungstenite::client::IntoClientRequest,
+    http::HeaderValue,
 };
 
 #[cfg(all(
@@ -1744,7 +1745,7 @@ async fn network_connect(
                 .into_client_request()?;
             request
                 .headers_mut()
-                .insert("Sec-WebSocket-Protocol", "mqtt".parse().unwrap());
+                .insert("Sec-WebSocket-Protocol", HeaderValue::from_static("mqtt"));
 
             if let Some(request_modifier) = options.fallible_request_modifier() {
                 request = request_modifier(request)
@@ -1776,7 +1777,7 @@ async fn network_connect(
                 .into_client_request()?;
             request
                 .headers_mut()
-                .insert("Sec-WebSocket-Protocol", "mqtt".parse().unwrap());
+                .insert("Sec-WebSocket-Protocol", HeaderValue::from_static("mqtt"));
 
             if let Some(request_modifier) = options.fallible_request_modifier() {
                 request = request_modifier(request)